+++ deployment/grid/terraform/compute_plane/variables.tf
@@ -169,25 +169,8 @@ variable "enable_private_subnet" {
   type        = bool
 }
 
-variable "grafana_configuration" {
-  description = "this variable store the configuration for the grafana helm chart"
-  type = object({
-    # downloadDashboardsImage_tag = string
-    # grafana_tag                 = string
-    # initChownData_tag           = string
-    # sidecar_tag                 = string
-    admin_password = string
-  })
+variable "grafana_admin_password" {
+  description = "Holds the default/initial password that will be used for authenticating with Grafana"
+  type        = string
+  sensitive   = true
 }
-
-# variable "prometheus_configuration" {
-#   description = "this variable store the configuration for the prometheus helm chart"
-#   type = object({
-#     node_exporter_tag      = string
-#     server_tag             = string
-#     alertmanager_tag       = string
-#     kube_state_metrics_tag = string
-#     pushgateway_tag        = string
-#     configmap_reload_tag   = string
-#   })
-# }